+++ nodes.h
@@ -15,12 +15,12 @@ class NodeS : public INodeS
 {
 public:
     SiteData _site;
-    float _rates[numberOfRates];
+    double _rates[numberOfRates];
 
     NodeS(int *cell, int **neighbours);
 
     void updateSum();
-    int reactionIndex(float r);
+    int reactionIndex(double r);
 };
 
 template <int numberOfRates>
@@ -30,14 +30,14 @@ NodeS<numberOfRates>::NodeS(int *cell, int **neighbours) : _site(cell, neighbour
 
 template <int numberOfRates>
 void NodeS<numberOfRates>::updateSum() {
-    float sum = 0;
+    double sum = 0;
     for (int i = 0; i < numberOfRates; ++i) sum += _rates[i];
     if (parent()) parent()->accSum(sum - _sum);
     _sum = sum;
 }
 
 template <int numberOfRates>
-int NodeS<numberOfRates>::reactionIndex(float r) {
+int NodeS<numberOfRates>::reactionIndex(double r) {
     for (int i = 0; i < numberOfRates - 1; ++i) {
         if (r < _rates[i]) return i;
         else r -= _rates[i];
