+++ rejectionfreesimulation_context.h
@@ -11,7 +11,7 @@ class RejectionFreeSimulationContext : public SimulationBaseContext
 public:
     RejectionFreeSimulationContext(AreaData *area);
 
-    float doReaction();
+    double doReaction();
 
 private:
     void reviewAllEvents();
@@ -19,7 +19,7 @@ private:
 private:
 //    std::list<EventData> _events;
     std::vector<EventData> _events;
-    float _totalRate;
+    double _totalRate;
 };
 
 #endif // REJECTIONFREESIMULATION_CONTEXT_H
